requirements.txt
Django==3.2.6
djangorestframework==3.12.4
django-cors-headers==3.8.0
Faker==8.12.0
numpy==1.21.2
